# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/riders": {
            "get": {
                "description": "gets all riders in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get all riders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Rider"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates a new rider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create rider",
                "parameters": [
                    {
                        "description": "Add rider",
                        "name": "rider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riderhdl.BodyCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/riderhdl.ResponseCreate"
                        }
                    }
                }
            }
        },
        "/riders/{id}": {
            "get": {
                "description": "gets a rider from the system by its ID",
                "produces": [
                    "application/json"
                ],
                "summary": "get rider",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rider id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Rider"
                        }
                    }
                }
            },
            "put": {
                "description": "updates a rider's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update rider",
                "parameters": [
                    {
                        "description": "Update rider",
                        "name": "rider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/riderhdl.BodyUpdate"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Rider id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/riderhdl.ResponseUpdate"
                        }
                    }
                }
            }
        },
        "/riders/{id}/location": {
            "put": {
                "description": "updates a rider's location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update rider location",
                "parameters": [
                    {
                        "description": "Update rider",
                        "name": "rider",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Location"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Rider id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/riderhdl.ResponseUpdate"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Location": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "domain.Rider": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/domain.Location"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "riderhdl.BodyCreate": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "riderhdl.BodyUpdate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "riderhdl.ResponseCreate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/domain.Location"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "riderhdl.ResponseUpdate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "location": {
                    "$ref": "#/definitions/domain.Location"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}